Create main
This code defines two functions: verify_eligibility() and verify_claim(). The verify_eligibility() function takes a member ID and a policy number as input, and sends a request to an API endpoint to verify the member's eligibility. The verify_claim() function takes a claim ID, a member ID, and a policy number as input, and sends a request to an API endpoint to verify the claim.

To use the code, you would need to replace the API endpoint URLs with the actual URLs of the API that you want to use. You would also need to make sure that you have the requests library installed, as it is used to send requests to the API.